# docs/06-api-and-integrations/artifacts/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "Body_upload_folder_repository_api_v1_repositories_upload_folder_post": {
        "properties": {
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "relative_paths": {
            "items": {
              "type": "string"
            },
            "title": "Relative Paths",
            "type": "array"
          }
        },
        "required": [
          "files",
          "relative_paths"
        ],
        "title": "Body_upload_folder_repository_api_v1_repositories_upload_folder_post",
        "type": "object"
      },
      "Body_upload_repository_api_v1_repositories_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_repository_api_v1_repositories_upload_post",
        "type": "object"
      },
...
      "CitationDTO": {
        "properties": {
          "end_line": {
            "title": "End Line",
            "type": "integer"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "index_version": {
            "default": 0,
            "title": "Index Version",
            "type": "integer"
          },
          "is_stale": {
            "default": false,
            "title": "Is Stale",
            "type": "boolean"
          },
          "start_line": {
            "title": "Start Line",
            "type": "integer"
          },
          "symbol_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol Name"
          }
        },
        "required": [
          "evidence_id",
          "file_path",
          "start_line",
          "end_line"
        ],
        "title": "CitationDTO",
        "type": "object"
      },
...
      "FileContentResponse": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "lines": {
            "items": {
              "type": "string"
            },
            "title": "Lines",
            "type": "array"
          },
          "symbols": {
            "items": {
              "$ref": "#/components/schemas/CitationDTO"
            },
            "title": "Symbols",
            "type": "array"
          }
        },
        "required": [
          "file_path",
          "language",
          "content",
          "lines"
        ],
        "title": "FileContentResponse",
        "type": "object"
      },
      "FileTreeNodeDTO": {
        "properties": {
          "children": {
            "items": {
              "$ref": "#/components/schemas/FileTreeNodeDTO"
            },
            "title": "Children",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "path",
          "type"
        ],
        "title": "FileTreeNodeDTO",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IgnorePatternsResponse": {
        "properties": {
          "default_patterns": {
            "items": {
              "type": "string"
            },
            "title": "Default Patterns",
            "type": "array"
          },
          "effective_patterns": {
            "items": {
              "type": "string"
            },
            "title": "Effective Patterns",
            "type": "array"
          },
          "user_patterns": {
            "items": {
              "type": "string"
            },
            "title": "User Patterns",
            "type": "array"
          }
        },
        "required": [
          "default_patterns",
          "effective_patterns"
        ],
        "title": "IgnorePatternsResponse",
        "type": "object"
      },
...
      "RepositoryBulkDeleteRequest": {
        "properties": {
          "delete_all": {
            "default": false,
            "title": "Delete All",
            "type": "boolean"
          },
          "repository_ids": {
            "items": {
              "type": "string"
            },
            "title": "Repository Ids",
            "type": "array"
          }
        },
        "title": "RepositoryBulkDeleteRequest",
        "type": "object"
      },
      "RepositoryBulkDeleteResponse": {
        "properties": {
          "deleted_count": {
            "title": "Deleted Count",
            "type": "integer"
          },
          "repository_ids": {
            "items": {
              "type": "string"
            },
            "title": "Repository Ids",
            "type": "array"
          }
        },
        "required": [
          "deleted_count"
        ],
        "title": "RepositoryBulkDeleteResponse",
        "type": "object"
      },
      "RepositoryCreateResponse": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "repository_id": {
            "title": "Repository Id",
            "type": "string"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "repository_id",
          "name",
          "status",
          "source_type"
        ],
        "title": "RepositoryCreateResponse",
        "type": "object"
      },
      "RepositoryDTO": {
        "properties": {
          "chunks": {
            "default": 0,
            "title": "Chunks",
            "type": "integer"
          },
          "current_index_version": {
            "default": 0,
            "title": "Current Index Version",
            "type": "integer"
          },
          "detected_stack": {
            "items": {
              "type": "string"
            },
            "title": "Detected Stack",
            "type": "array"
          },
          "endpoints": {
            "default": 0,
            "title": "Endpoints",
            "type": "integer"
          },
          "graph_nodes": {
            "default": 0,
            "title": "Graph Nodes",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "indexed_files": {
            "default": 0,
            "title": "Indexed Files",
            "type": "integer"
          },
          "last_indexed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Indexed At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Label"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "source_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Uri"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "symbols": {
            "default": 0,
            "title": "Symbols",
            "type": "integer"
          },
          "total_files": {
            "default": 0,
            "title": "Total Files",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "source_type",
          "status"
        ],
        "title": "RepositoryDTO",
        "type": "object"
      },
      "RepositoryDeleteResponse": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "repository_id": {
            "title": "Repository Id",
            "type": "string"
          }
        },
        "required": [
          "deleted",
          "repository_id"
        ],
        "title": "RepositoryDeleteResponse",
        "type": "object"
      },
...
      "SearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResultDTO"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
...
      "SettingsResponse": {
        "properties": {
          "indexing": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "title": "Indexing",
            "type": "object"
          },
          "providers": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "title": "Providers",
            "type": "object"
          },
          "security": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Security",
            "type": "object"
          }
        },
        "required": [
          "indexing",
          "providers",
          "security"
        ],
        "title": "SettingsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AI Codebase Assistant",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/repositories": {
      "get": {
        "operationId": "list_repositories_api_v1_repositories_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RepositoryDTO"
                  },
                  "title": "Response List Repositories Api V1 Repositories Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Repositories",
        "tags": [
          "repositories"
        ]
      }
    },
    "/api/v1/repositories/bulk-delete": {
      "post": {
        "operationId": "delete_repositories_api_v1_repositories_bulk_delete_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryBulkDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryBulkDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Repositories",
        "tags": [
          "repositories"
        ]
      }
    },
    "/api/v1/repositories/upload": {
      "post": {
        "operationId": "upload_repository_api_v1_repositories_upload_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_repository_api_v1_repositories_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Repository",
        "tags": [
          "repositories"
        ]
      }
    },
    "/api/v1/repositories/upload-folder": {
      "post": {
        "operationId": "upload_folder_repository_api_v1_repositories_upload_folder_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_folder_repository_api_v1_repositories_upload_folder_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Folder Repository",
        "tags": [
          "repositories"
        ]
      }
    },
    "/api/v1/repositories/{repository_id}": {
      "delete": {
        "operationId": "delete_repository_api_v1_repositories__repository_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Repository",
        "tags": [
          "repositories"
        ]
      }
    },
...
    "/api/v1/repositories/{repository_id}/files/content": {
      "get": {
        "operationId": "get_file_content_api_v1_repositories__repository_id__files_content_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get File Content",
        "tags": [
          "repositories"
        ]
      }
    },
    "/api/v1/repositories/{repository_id}/files/tree": {
      "get": {
        "operationId": "get_file_tree_api_v1_repositories__repository_id__files_tree_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FileTreeNodeDTO"
                  },
                  "title": "Response Get File Tree Api V1 Repositories  Repository Id  Files Tree Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get File Tree",
        "tags": [
          "repositories"
        ]
      }
    },
...
    "/api/v1/repositories/{repository_id}/search": {
      "get": {
        "operationId": "search_api_v1_repositories__repository_id__search_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "repositories"
        ]
      }
    },
...
    "/api/v1/settings": {
      "get": {
        "operationId": "get_settings_api_v1_settings_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Settings",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/ignore-patterns": {
      "get": {
        "operationId": "get_ignore_patterns_api_v1_settings_ignore_patterns_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IgnorePatternsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ignore Patterns",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}